making requested changes
